requirements.txt
functions-framework==3.*
Flask
flask-cors
google-cloud-firestore
google-cloud-aiplatform
vertexai